app/auth/jwt.py
from jose import jwt
from datetime import datetime, timedelta
from config import settings

def create_access_token(data: dict) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(
        minutes=settings.JWT_EXPIRE_MINUTES
    )
    to_encode.update({"exp": expire})
    return jwt.encode(
        to_encode,
        settings.JWT_SECRET,
        algorithm=settings.JWT_ALGORITHM
    )

def verify_token(token: str) -> dict:
    try:
        payload = jwt.decode(
            token,
            settings.JWT_SECRET,
            algorithms=[settings.JWT_ALGORITHM]
        )
        if payload.get("exp") < datetime.utcnow().timestamp():
            raise ValueError("Token expired")
        return payload
    except Exception as e:
        raise ValueError(f"Invalid token: {str(e)}")
