requirements.txt
annotated-types==0.7.0
asyncpg==0.27.0
Authlib==1.2.0
cryptography==44.0.3
cffi==1.17.1
pycparser==2.22
bcrypt==4.0.1
fastapi==0.103.1
anyio==3.7.1
idna==3.10
sniffio==1.3.1
pydantic==1.10.7
typing_extensions==4.13.2
starlette==0.27.0
httpx==0.24.1
certifi==2025.4.26
httpcore==0.17.3
h11==0.14.0
passlib==1.7.4
pipdeptree==2.26.1
packaging==25.0
pip==25.1.1
psycopg2-binary==2.9.6
pydantic_core==2.33.2
pytest==8.3.5
colorama==0.4.6
iniconfig==2.1.0
pluggy==1.6.0
python-dateutil==2.8.2
six==1.17.0
python-dotenv==1.0.0
python-jose==3.3.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1
python-multipart==0.0.6
setuptools==80.7.1
SQLAlchemy==2.0.15
greenlet==3.2.2
typing-inspection==0.4.0
uvicorn==0.22.0
click==8.2.0
wheel==0.45.1
